requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
beautifulsoup4==4.11.2
certifi==2022.12.7
charset-normalizer==3.1.0
dnspython==2.3.0
fake-useragent==1.1.1
flake8==6.0.0
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
isort==5.12.0
lxml==4.9.2
magic-filter==1.0.9
mccabe==0.7.0
mongoengine==0.27.0
multidict==6.0.4
psycopg2==2.9.5
pycodestyle==2.10.0
pyflakes==3.0.1
pymongo==4.3.3
python-dotenv==1.0.0
pytz==2022.7.1
requests==2.28.2
soupsieve==2.4
SQLAlchemy==2.0.6
typing_extensions==4.5.0
urllib3==1.26.15
yarl==1.8.2
